Phase 2 step 2: Extract pure render cells + text_blocks + charts
